syscallcompat: Use the same *User() syscall wrappers on both Linux and macOS.
